backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
